refactor Merge.py (#16)
