+++ udf/README.md
@@ -1,0 +1,6 @@
+## Deployment
+
+- copy `api` into `${IGINX_HOME}/udf_funcs/python_scripts/`
+- copy `default` into `${IGINX_HOME}/udf_funcs/python_scripts/`
+- copy `udf_list` into `${IGINX_HOME}/udf_funcs/`
+
